distributing parallel module
